処理を整理
