+++ src/main.ts
@@ -1,63 +1,16 @@
 import {Message, Client} from 'discord.js';
-import {Octokit} from 'octokit';
-
 import dotenv from 'dotenv';
-import express from 'express';
+
+import {createDummyServer} from './dummyServer';
+import {fetchUserData, inviteUser} from './githubCommands';
 
 dotenv.config();
 
-// GAEがサーバーをリスンしていないとそもそもアプリとして認識してくれないので、ダミーのレスポンスを返すようにしておく
-const PORT = Number(parseInt(`${process.env.PORT}`)) || 8080;
-const app = express();
-
-app.get('/', (_req, res) => {
-  res.send('🤖Bot is running!!🤖');
-});
-
-export const server = app.listen(PORT, () => {
-  console.log(`App listening on port ${PORT}`);
-});
-
-const octokit = new Octokit({auth: process.env.GITHUB_PAT});
+createDummyServer();
 
 const client = new Client({
   intents: ['GUILDS', 'GUILD_MEMBERS', 'GUILD_MESSAGES'],
 });
-
-const fetchUserData = async (userName: string) => {
-  try {
-    const invitedUserResponse = await octokit.request(
-        'GET /users/{username}', {
-          username: userName,
-        },
-    );
-
-    if (invitedUserResponse.status != 200) {
-      console.log(`${userName} is not valid user`);
-      return null;
-    }
-
-    return invitedUserResponse.data;
-  } catch (err) {
-    console.log(err);
-    return null;
-  }
-};
-
-const inviteUser = async (userId: number) => {
-  try {
-    const invitationResponse = await octokit.request(
-        'POST /orgs/{org}/invitations', {
-          org: 'yurugengo-supporters',
-          invitee_id: userId,
-        });
-
-    return invitationResponse;
-  } catch (err) {
-    console.log(err);
-    return null;
-  }
-};
 
 const githubCommandName = 'ghinvite';
